pyproject.toml
[project]
name = "kartli"
version = "0.1.0"
description = "Generate static map images with Swisstopo and OpenStreetMap tiles"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "Pillow>=11.0",
    "httpx>=0.28",
]

[project.scripts]
kartli = "kartli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kartli"]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
    "pytest>=9.0.2",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "S", "C90"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]
